mrc-moe-ep-setup(extract-quant-moe-checks-move,non_mechanical_provable): Move check_quantized_moe_compatibility and its _use_aiter constant to moe_ep_setup module
PR-Title: Extract MoE/EP setup into a moe_ep_setup module
